% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_MODIS.R
\name{downloadMODIS}
\alias{downloadMODIS}
\title{downloadMODIS}
\usage{
downloadMODIS(
  study_area_boundaries,
  date,
  satellite,
  username,
  password,
  out_path,
  buff = 5
)
}
\arguments{
\item{study_area_boundaries}{"SpatialPolygonsDataFrame", "SpatVector", "SpatialPolygons". Boundaries of the study area}

\item{date}{"Date", "POSIXlt", "POSIXct". Date to download}

\item{satellite}{"character". Satellite to download. One of:
\itemize{
\item "MYD10A1" - Download Aqua
\item"MOD10A1" - Doownload Terra
\item"Combined" - Combination of Terra and Aqua products. Terra missing values are filled with Aqua
}}

\item{username}{"character". MODIS download username aunthetification}

\item{password}{"character". MODIS download password aunthetification}

\item{out_path}{"character". Output directory, if out_path do not exist it will be created}

\item{buff}{"integrer". Number off cells included as buffer in the procesing}
}
\description{
Download and preprocess Snow Cover Daily MODIS products. The outputs will be cloud masked NDSI products and stored in the provided directory. If the file exists it will not be overwritten.
}
\details{
Highly modified from the 'MODISSnow' package
}
\examples{
\dontrun{
# Load a shapefile
data("Pineta")

#create temporal dir to store the data
out_path = tempdir()

# Download MODIS NDSI product (use your own credentials)
downloadMODIS(study_area_boundaries = Pineta,
              date = as.Date("2015-06-01"),
              satellite  = "Combined",
              username = "user",
              password = "pass",
              out_path = out_path)

#List files in out_path
mod_files = list.files(out_path,full.names = T)
}

}
