% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_S2_hub.R
\name{downloadS2}
\alias{downloadS2}
\title{downloadS2
Not implemented yet}
\usage{
downloadS2(
  study_area_boundaries,
  time_window,
  username,
  password,
  max_cloud,
  out_path,
  do1C,
  tiles = c("ALL"),
  LTAorder = T,
  buff = 10,
  parallel = F,
  rm_summer = T,
  clean = T
)
}
\description{
downloadS2
Not implemented yet
}
