% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_S2_theia.R
\name{downloadS2_theiasnow}
\alias{downloadS2_theiasnow}
\title{downloadS2_theiasnow}
\usage{
downloadS2_theiasnow(
  study_area_boundaries,
  time_window,
  max_cloud,
  username,
  password,
  out_path,
  buff = 10,
  tile = NULL,
  avoid = NULL
)
}
\arguments{
\item{study_area_boundaries}{"SpatialPolygonsDataFrame", "SpatVector", "SpatialPolygons". Boundaries of the study area}

\item{time_window}{"Date", "POSIXlt", "POSIXct". Two date vector storing the begining and end of the time window.}

\item{max_cloud}{"integrer", Maximun percentage of clouds allowed form 0 to 100 percentage.}

\item{username}{"character". Theia download username aunthetification}

\item{password}{"character". Theia download password aunthetification}

\item{out_path}{"character". Output directory, if out_path do not exist it will be created}

\item{buff}{"integrer". Number off cells included as buffer in the procesing}

\item{tile}{"character". Optional. Tile to download}

\item{avoid}{"integrer". Optional. Numeric vector with the months that should not be downlaoded.}
}
\description{
Download and preprocess Theia snow cover products products. The outputs will be cloud masked binary snow cover stored in the provided directory. If the file exists it will not be overwritten.
}
\details{
This interim version uses system calls to the [theia_download script](https://github.com/ealonsogzl/theia_download)
}
\examples{
\dontrun{
# Load a shapefile
data("Pineta")

#create temporal dir to store the data
out_path = tempdir()

# Download Theia snow cover products (use your own credentials)
downloadS2_theiasnow(study_area_boundaries = Pineta,
              time_window = c(as.Date("2015-08-01"),Sys.Date()),
              max_cloud = 30,
              username = "user",
              password = "pass",
              out_path = out_path,
              avoid = c(7,8) #Avoid to download July and Agust
              )

#List files in out_path
s2_files = list.files(out_path, full.names = T)
}

}
