% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_MODIS.R
\name{modissnow_get_data}
\alias{modissnow_get_data}
\title{modissnow_get_data
Internal function. download MODIS scene}
\usage{
modissnow_get_data(date, sat, h, v, user, passwd, clean = T)
}
\arguments{
\item{date}{date to request}

\item{sat}{satellite}

\item{h}{h coordinate}

\item{v}{v cordinate}

\item{user}{username}

\item{passwd}{user password}

\item{clean}{Clean the temp filesystem}
}
\description{
modissnow_get_data
Internal function. download MODIS scene
}
\keyword{internal}
