% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_S2_theia.R
\name{s2_theia_get_data}
\alias{s2_theia_get_data}
\title{s2_theia_get_data
Internal function. Setup theia downloads}
\usage{
s2_theia_get_data(
  study_area_boundaries,
  time_window,
  out_path,
  max_cloud,
  theia_instr,
  buff,
  tile = NULL,
  avoid = NULL,
  product = "Snow",
  clean = T
)
}
\arguments{
\item{study_area_boundaries}{username}

\item{time_window}{password}

\item{out_path}{username}

\item{max_cloud}{password}

\item{theia_instr}{username}

\item{buff}{password}

\item{tile}{password}

\item{avoid}{password}

\item{clean}{}
}
\description{
s2_theia_get_data
Internal function. Setup theia downloads
}
\keyword{internal}
