% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_S2_theia.R
\name{setup_theia}
\alias{setup_theia}
\title{setup_theia
Internal function. Setup theia downloads}
\usage{
setup_theia(username, password)
}
\arguments{
\item{username}{username}

\item{password}{password}
}
\description{
setup_theia
Internal function. Setup theia downloads
}
\keyword{internal}
