% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_MODIS.R
\name{tidy_modis}
\alias{tidy_modis}
\title{tidy_modis
Internal function. Clean modis scenes}
\usage{
tidy_modis(modis_tile, name_out)
}
\arguments{
\item{modis_tile}{raster modis file}

\item{name_out}{name of the file}
}
\description{
tidy_modis
Internal function. Clean modis scenes
}
\keyword{internal}
